refactor: extract AuthPageFactory from WebPageFactory
Moves 6 auth methods (~131 lines) into a dedicated AuthPageFactory
class with zero constructor dependencies. WebPageFactory delegates
to it, keeping full backward compatibility. Follows the same pattern
as the existing AdminPageFactory delegation.
